IT_Chat_Bot/Knowledge_Base/requirements.txt
fastapi
uvicorn==0.29.0
transformers
torch==2.3.0
python-multipart
PyMuPDF
python-docx
pydantic
docs
typing

